requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
cryptography==37.0.4
pycparser==2.21
